+++ java-client-sdk/lib/src/main/java/org/openapis/openapi/SDKConfiguration.java
@@ -30,8 +30,8 @@ class SDKConfiguration {
 	}};
 	public String language = "java";
 	public String openapiDocVersion = "0.1.0";
-	public String sdkVersion = "1.27.1";
-	public String genVersion = "2.86.6";
+	public String sdkVersion = "1.27.2";
+	public String genVersion = "2.87.1";
 	public Map<String, Map<String, Map<String, Object>>> globals = new HashMap<String, Map<String, Map<String, Object>>>(){{
 		put("parameters", new HashMap<String, Map<String, Object>>());
 	}};